PR libfortran/27107 	* runtime/environ.c: Don't include io/io.h. 	* runtime/string.c: Don't include io/io.h. 	(compare0): Add cast to avoid warning. 	* runtime/error.c: Don't include io/io.h. 	(st_printf): Move to io/unix.c. 	* intrinsics/flush.c: Delete, contents moved to io/intrinsics.c. 	* intrinsics/fget.c: Likewise. 	* intrinsics/ftell.c: Likewise. 	* intrinsics/tty.c: Likewise. 	* libgfortran.h (DEFAULT_RECL, notification_std, 	get_unformatted_convert, IOPARM_*, st_parameter_common, unit_convert, 	DEFAULT_TEMPDIR): New declarations. 	* io/io.h (DEFAULT_RECL, notification_std, get_unformatted_convert, 	IOPARM_*, st_parameter_common, unit_convert, DEFAULT_TEMPDIR): 	Move to libgfortran.h. 	* io/unix.c: Add io/unix.h content. 	(st_printf): New function. 	* io/intrinsics.c: New file. 	* io/unix.h: Remove, contents moved into unix.c. 	* libtool-version: Update library version to 3.0.0. 	* configure.ac: Update library version to 0.3. 	* Makefile.am (intrinsics/fget.c, intrinsics/flush.c, 	intrinsics/ftell.c, intrinsics/tty.c, libgfortran.h): Remove targets. 	* Makefile.in: Regenerate. 	* configure: Regenerate.
git-svn-id: svn+ssh://gcc.gnu.org/svn/gcc/trunk@120869 138bc75d-0d04-0410-961f-82ee72b054a4

+++ libgfortran/runtime/string.c
@@ -31,7 +31,6 @@ Boston, MA 02110-1301, USA.  */
 #include <string.h>
 
 #include "libgfortran.h"
-#include "../io/io.h"
 
 /* Compare a C-style string with a fortran style string in a case-insensitive
    manner.  Used for decoding string options to various statements.  Returns
@@ -44,7 +43,7 @@ compare0 (const char *s1, int s1_len, const char *s2)
 
   /* Strip trailing blanks from the Fortran string.  */
   len = fstrlen (s1, s1_len);
-  if(len != strlen(s2)) return 0; /* don't match */
+  if (len != (int) strlen(s2)) return 0; /* don't match */
   return strncasecmp (s1, s2, len) == 0;
 }
